fix(logging): Correctly determine release type in CleanLogger
The `CleanLogger` was incorrectly determining the release type because the `name` variable was not being passed to `magic.is_dev_release` and `magic.is_pre_release`. This commit fixes the issue by removing the re-assignment of `name` within the `CleanLogger` constructor, allowing the correct `name` from the frame's globals to be used.